requirements.txt
tensorflow==2.13.0
numpy>=1.19.2
opencv-python>=4.5.1
Flask==2.0.1
Flask-CORS==3.0.10
Pillow>=8.2.0
pandas>=1.2.4
scikit-learn>=0.24.2
Werkzeug==2.0.3 